fix: address review feedback for backend security hardening
- Restore original absolute path handling in environment.py (allow
  absolute paths within env_base_dir boundary)
- Add _task_locks_mutex to get_task_lock, get_task_lock_if_exists,
  and delete_task_lock for thread safety
- Merge mutex tests from test_task_lock_mutex.py into test_task.py
  using def test_xxx style
- Update test_environment.py to match restored path behavior
